+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Bug fixes:
 * Calling `Kernel#raise` with a raised exception will no longer set the cause of the exception to itself (#1682).
 * Return a `FFI::Function` correctly for functions returning a callback.
 * Convert to intuitive Ruby exceptions when INVOKE fails (#1690).
+* Implemented `FFI::Pointer#clear` (#1687).
 
 Compatibility